Move ParseURLHost() to httputils
